+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Interactive demo
+
+### Changed
+
+- Separated language pattern from annotations when rendering score
+- Integrated `promise-worker@2.0.1` in worker-scripts
+
 ## [1.3.3] - 2024-03-28
 
 ### Changed
@@ -17,7 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bumped dependencies, Node-requirements
 - Simplified README
-- Prepare API-move
+- Prepare /docs/api move
 
 ### Added